+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Geld in info app afronden voor komma getal verbergen
 - Tijdelijk disabled dat stalling de voertuigservice beinvloed tot ik service heb gefixt
 - Je stapt niet langer soms automatisch uit na het lockpicken
+- Clamped vehicle service values
 
 ## [1.0.1]